requirements.txt
amqp==5.3.1
asgiref==3.6.0
billiard==4.2.2
black==23.3.0
celery==5.5.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==40.0.2
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2
django-allauth==0.54.0
django-cors-headers==4.9.0
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-filter==23.2
django-htmx==1.26.0
django-import-export==4.3.14
django-nested-admin==4.0.2
django-template-partials==25.3
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==18.6.0
fonttools==4.60.0
idna==3.4
iniconfig==2.3.0
kiwisolver==1.4.9
kombu==5.5.4
matplotlib==3.10.6
mypy-extensions==1.0.0
numpy==2.3.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==23.1
pandas==2.3.2
pathspec==0.11.1
pillow==11.3.0
platformdirs==3.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.21
Pygments==2.19.2
PyJWT==2.6.0
pyparsing==3.2.5
pytesseract==0.3.13
pytest==9.0.0
pytest-django==4.11.1
python-dateutil==2.8.2
python-dotenv==1.2.1
python-monkey-business==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.29.0
requests-oauthlib==1.3.1
ruff==0.13.1
seaborn==0.13.2
six==1.16.0
sqlparse==0.4.4
tablib==3.9.0
tabulate==0.9.0
tomli==2.0.1
tzdata==2025.2
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.14
xlrd==2.0.2
xlwt==1.3.0
